# dist/swagger_output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "v1.0.0",
        "title": "Transaction and User API",
        "description": "A simple API to simulate simple transactions and user management"
    },
    "servers": [
        {
            "url": "http://localhost:4000/api",
            "description": "Development"
        },
        {
            "url": "https://txn-api.vercel.app/api",
            "description": "Production"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "summary": "Create a new user",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The user was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "summary": "User login",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "summary": "Get user profile",
                "tags": ["Profile"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "User profile fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "summary": "Update user profile",
                "tags": ["Profile"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "firstName": { "type": "string" },
                                    "lastName": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Delete user profile",
                "tags": ["Profile"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "User profile deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "summary": "Create a new transaction",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_id": { "type": "string" },
                                    "amount": { "type": "number" },
                                    "transaction_type": { "type": "string" },
                                    "description": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The transaction was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "get": {
                "summary": "Returns the list of all transactions for the authenticated user",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": { "type": "integer" },
                        "description": "The page number for pagination"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": { "type": "integer" },
                        "description": "The number of items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The list of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "summary": "Get the transaction by id",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": { "type": "string" },
                        "required": true,
                        "description": "The transaction id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The transaction description by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The transaction was not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "summary": "Update the transaction by id",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": { "type": "string" },
                        "required": true,
                        "description": "The transaction id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "amount": { "type": "number" },
                                    "transaction_type": { "type": "string" },
                                    "description": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The transaction was successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The transaction was not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "summary": "Delete the transaction by id",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": { "type": "string" },
                        "required": true,
                        "description": "The transaction id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The transaction was successfully deleted"
                    },
                    "404": {
                        "description": "The transaction was not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/transactions-summary": {
            "get": {
                "summary": "Get a summary of transactions (total credits, total debits, balance) for the authenticated user",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "The summary of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "totalCredits": { "type": "number" },
                                        "totalDebits": { "type": "number" },
                                        "balance": { "type": "number" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/historical-transaction-volume": {
            "get": {
                "summary": "Get historical chart data for transaction volume",
                "tags": ["Transactions"],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "The historical chart data for transaction volume",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "dates": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "volumes": {
                                            "type": "array",
                                            "items": { "type": "number" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": { "type": "string" },
                    "user_id": { "type": "string" },
                    "amount": { "type": "number" },
                    "transaction_type": { "type": "string" },
                    "description": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": { "type": "string" },
                    "email": { "type": "string" },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}